Update deprecated media constants
The `MEDIA_CLASS_*` and `MEDIA_TYPE_*` constants are deprecated and produce warn logs like these:

`MEDIA_TYPE_ARTIST was used from airmusic, this is a deprecated constant which will be removed in HA Core 2025.10. Use MediaType.ARTIST instead, please report it to the author of the 'airmusic' custom integration`

I replaced the usage of them with the new constants and updated the imports accordingly.